rename(manual): 폴더 구조 개선
