+++ src/app/(route)/manual/_components/MANUAL_CONST.tsx
